requirements.txt
torch>=2.0
torchvision>=0.15
numpy<2
opencv-python-headless>=4.9
Pillow>=10.0
scikit-learn>=1.3
tqdm>=4.60
requests>=2.31
lightglue @ git+https://github.com/cvg/LightGlue.git
